move the sha1 functions to the library
