+++ lib/lib_sha1.c
@@ -2,25 +2,35 @@
  * @file
  * Calculate the SHA1 checksum of a buffer
  *
- * SHA-1 in C
+ * @authors
+ * Steve Reid <steve@edmweb.com>
+ * Thomas Roessler <roessler@does-not-exist.org>
  *
- * By Steve Reid <steve@edmweb.com>, with small changes to make it
- * fit into mutt by Thomas Roessler <roessler@does-not-exist.org>.
+ * @copyright
+ * Public Domain
+ */
+
+/**
+ * @page sha1 Calculate the SHA1 checksum of a buffer
  *
- * 100% Public Domain.
+ * Calculate the SHA1 cryptographic hash of a string, according to RFC3174.
  *
- * Test Vectors (from FIPS PUB 180-1)
- * "abc"
- * A9993E36 4706816A BA3E2571 7850C26C 9CD0D89D
- * "abcdbcdecdefdefgefghfghighijhijkijkljklmklmnlmnomnopnopq"
- * 84983E44 1C3BD26E BAAE4AA1 F95129E5 E54670F1
- * A million repetitions of "a"
- * 34AA973C D4C4DAA4 F61EEB2B DBAD2731 6534016F
+ * | Function         | Description
+ * | :--------------- | :----------------------------------------
+ * | sha1_final()     | Add padding and return the message digest
+ * | sha1_init()      | Initialize new context
+ * | sha1_transform() | Hash a single 512-bit block
+ * | sha1_update()    | Run your data through this
+ *
+ * Test Vectors (from FIPS PUB 180-1):
+ * - "abc" yields `A9993E36 4706816A BA3E2571 7850C26C 9CD0D89D`
+ * - "abcdbcdecdefdefgefghfghighijhijkijkljklmklmnlmnomnopnopq" yields `84983E44 1C3BD26E BAAE4AA1 F95129E5 E54670F1`
+ * - A million repetitions of "a" yields `34AA973C D4C4DAA4 F61EEB2B DBAD2731 6534016F`
  */
 
 #include "config.h"
 #include <string.h>
-#include "sha1.h"
+#include "lib_sha1.h"
 
 #define rol(value, bits) (((value) << (bits)) | ((value) >> (32 - (bits))))
 
@@ -55,9 +65,10 @@
   z += (w ^ x ^ y) + blk(i) + 0xCA62C1D6 + rol(v, 5);                          \
   w = rol(w, 30);
 
-
 /**
  * sha1_transform - Hash a single 512-bit block
+ * @param state  Internal state of the transform
+ * @param buffer Data to transform
  *
  * This is the core of the algorithm.
  */
@@ -168,9 +179,9 @@ void sha1_transform(uint32_t state[5], const unsigned char buffer[64])
   memset(block, '\0', sizeof(block));
 }
 
-
 /**
  * sha1_init - Initialize new context
+ * @param context SHA1 context
  */
 void sha1_init(struct Sha1Ctx *context)
 {
@@ -183,9 +194,11 @@ void sha1_init(struct Sha1Ctx *context)
   context->count[0] = context->count[1] = 0;
 }
 
-
 /**
  * sha1_update - Run your data through this
+ * @param context SHA1 context
+ * @param data    Data to be hashed
+ * @param len     Length of data
  */
 void sha1_update(struct Sha1Ctx *context, const unsigned char *data, uint32_t len)
 {
@@ -212,9 +225,10 @@ void sha1_update(struct Sha1Ctx *context, const unsigned char *data, uint32_t le
   memcpy(&context->buffer[j], &data[i], len - i);
 }
 
-
 /**
  * sha1_final - Add padding and return the message digest
+ * @param[out] digest  Message digest (SHA1 sum)
+ * @param[in]  context SHA1 context
  */
 void sha1_final(unsigned char digest[20], struct Sha1Ctx *context)
 {
