+++ lib/lib_sha1.h
@@ -1,8 +1,10 @@
 /**
  * @file
- * Convenience wrapper for the library headers
+ * Calculate the SHA1 checksum of a buffer
  *
  * @authors
+ * Copyright (C) 2000 Steve Reid <steve@edmweb.com>
+ * Copyright (C) 2000 Thomas Roessler <roessler@does-not-exist.org>
  * Copyright (C) 2017 Richard Russon <rich@flatcap.org>
  *
  * @copyright
@@ -20,30 +22,26 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#ifndef _LIB_SHA1_H
+#define _LIB_SHA1_H
+
+#include <stdint.h>
+
 /**
- * @page lib Library of shared functions
- *
- * Each source file in the library provides a group of related functions.
- *
- * The library is self-contained -- some files may depend on others in the
- * library, but none depends on source from outside.
- *
- * -# @subpage ascii
- * -# @subpage base64
- * -# @subpage date
- * -# @subpage exit
- * -# @subpage md5
- * -# @subpage sha1
+ * struct Sha1Ctx - Cursor for the SHA1 hashing
  */
+struct Sha1Ctx
+{
+  uint32_t state[5];
+  uint32_t count[2];
+  unsigned char buffer[64];
+};
 
-#ifndef _LIB_LIB_H
-#define _LIB_LIB_H
+void sha1_final(unsigned char digest[20], struct Sha1Ctx *context);
+void sha1_init(struct Sha1Ctx *context);
+void sha1_transform(uint32_t state[5], const unsigned char buffer[64]);
+void sha1_update(struct Sha1Ctx *context, const unsigned char *data, uint32_t len);
 
-#include "lib_ascii.h"
-#include "lib_base64.h"
-#include "lib_date.h"
-#include "lib_exit.h"
-#include "lib_md5.h"
-#include "lib_sha1.h"
+#define SHA_DIGEST_LENGTH 20
 
-#endif /* _LIB_LIB_H */
+#endif /* _LIB_SHA1_H */
